drop red/green diffstat from manpage version list
The dropdown box that shows which versions of each manpage are available
also figures out which versions actually modified that manpage (versus
the previous one). This is helpful to omit versions which are identical.

But it also shows a little diffstat image with the number of
added/removed lines. This is cool to look at, but not really all that
useful in practice (you might get a sense if there were a lot of changes
or just a few, but more likely you just want to go to the version that
matches what you have installed locally).

Worse, this diffstat means that an uncached request has to pull down
_every_ version of the manpage and do O(n) diffs. This can sometimes
cause timeouts for large pages that see frequent modifications (e.g.,
`git-config` is probably the worst offender; it topped 15s recently,
causing server errors in https://github.com/git/git-scm.com/issues/1563).

So let's drop the diffstat entirely. This drops the page-generation time
for docs/git-config on my laptop from 5000ms to 150ms.

This also lets us drop some css and ruby code, as well as the dependency
on the diff-lcs gem.