guis.yml: reorder yaml to match original form
The order fields were updated in d79886aa (Reorder Git GUI list
according to popularity, 2021-01-20), but that commit also reordered
the yaml file itself. This makes the diff quite hard to follow.

I restored the original order with a little perl magic:

  git cat-file blob d79886aa^:resources/gui.yml >old.yml
  git cat-file blob d79886aa:resources/gui.yml >new.yml
  perl -MYAML -e '
    $YAML::Preserve = 1;
    $YAML::Indent = 0;

    my %orders;
    my $new = YAML::LoadFile("new.yml");
    for my $gui (@{$new->{guis}}) {
      $orders{$gui->{name}} = $gui->{order};
    }

    my $old = YAML::LoadFile("old.yml");
    for my $gui (@{$old->{guis}}) {
      $gui->{order} = $orders{$gui->{name}};
    }

    print YAML::Dump($old);
  ' >resources/gui.yml

Unfortunately our quoting and indentation are not entirely consistent,
so I had to pick apart the useful changes with:

  git checkout -p d79886aa^ resources/gui.yml

And this is the result. The diff here is likewise messy and unreadable,
but when diffed against d79886aa^ (which is what will show in PR #1579),
it shows the overall change in the simplest way.