rubocop: disable InsecureHashAlgorithm check
We use sha1 as a checksum for deciding when pages are out of date (so a
weak algorithm is fine). This is really independent of Git's use of
sha1, so I think it could just be changed. But it would probably involve
a complete re-import of all of the pages. So let's just ignore it for
the purposes of this PR.