Add Fugitive GUI and sort the GUI list
Also add a screenshot of the user interface.

Sort the GUI order by Google Trends using the sort-gui script. Add some
trend_name attributes to GUIs that have confusing names that may be
increasing their popularity. Use a keyword like "Git" or "client" along
with the software's name to try to disambiguate it.