stylesheets: improve fonts for Linux
In Linux systems typically the first options don't exist, so the page
ends up with 'serif', which is how hello world pages look like.

At the very least we should specify a readable option: serif.

Also, DejaVu Serif is good alternative for Georgia.

And Roboto Slab is very similar to Adelle.

This makes it so the fonts are readable in Linux systems.

Other systems should not be affected.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>