book2: match ebook release assets by name, not content-type
Since progit2 2.1.309 switched to using a GitHub Action to generate
releases, it no longer sets the content-type field of the release
assets. And thus we fail to find any ebooks to link, since we are
looking for assets with "application/pdf", and so on.

Instead, let's do a suffix match against the name of each asset, looking
for the appropriate file extension. This works fine, since the names are
all obvious ("progit.epub", and so on). And it should be
backwards-compatible with older translations that haven't yet picked up
the new Actions-based workflow, since they used the same sensible names.

We may revert this later if the assets start generating with correct
content-types again. But it looks to be non-trivial (there's some
discussion in #1498), so this seems like a good solution in the
meantime.