ci: only run ci on pull request
We trigger CI on both "push" and "pull_request". The latter is necessary
to get CI results for people opening a PR from their own fork. The
former is not strictly necessary, but it's nice to get the PR running as
soon as we know about the commits, so that it's often finished by the
time you've finished opening up the PR.

But this does result in duplicate runs for branches pushed directly into
the repo: once when the branch is pushed, and once when the PR is
opened. We could get around this by skipping the workflow when we've
already got the same job running, but that's more complicated. Let's
just drop the "push" trigger. The CI here is pretty fast anyway (at
least once the bundler bits are all cached, but that should be the case
for any PR that doesn't update gems).