Gemfile: drop launchy and netrc gems
These were added in 5b270587 (add rest client gem, 2012-04-04), but we
later dropped the rest-client gem in c4104694 (Rails 3.2.17,
2014-04-27). It's not clear to me if they were needed as part of that,
or if it was simply an error (they actually got moved out of the dev
section). But they don't seem to be doing anything in the modern
codebase. Let's get rid of them in the name of simplicity.