app/vites/site/about: add FusionAuth as a sponsor
Per [1], Dan Moore of FusionAuth has graciously offered (and the Git
project accepted) to make an annual donation to the Git project which
is used to cover our Heroku costs to run git-scm.com.

Include a link to their company's website in the "sponsors" section of
the /site/about page, as well as a small blurb about their donation.

Thanks, FusionAuth!

[1]: https://lore.kernel.org/git/CAKUhyqHSwUb3UL=sKxqNrZsRNbf=wuzsk06LFfE=UwdN4JN37g@mail.gmail.com/