book: mention that we are an updated version of the 2nd edition
The intro on the main book page doesn't make it clear that this is the
2nd edition plus many years of updates. Let's say so.

I initially thought I'd rework the first paragraph, but I think
everything it says is still accurate and something we want to be saying.
So I ended up just tacking on an extra note.

The link to the progit repository is a little redundant, since we have
one in the sidebar already. But it's easy to miss, and I think makes
sense to emphasize here in the book overview.