update to ruby 3.1.2
The current version we specify is 2.7.3, but it no longer builds on
Debian unstable due to openssl 1.1 being too recent (and removing and
deprecating several interfaces).

Let's just move on to 3.1.2, which is the current stable version. It's a
big jump, but we don't really do anything too exotic in our code base,
so there's a good chance it's fine (which we can verify through
inspection).