replace rubocop-github gem with vanilla rubocop
It doesn't make much sense to use the -github variant, which is designed
for GitHub-specific projects (which we are not). Moreover, they depend
on an old version of rubocop which doesn't know about our more recent
ruby version.

So let's just swap it out for the regular rubocop gem. We'll need to add
the -rails and -performance dependencies ourselves. And while we're
here, we'll bump the targer ruby version in the rubocop file.