book: remove git-credential-read-only import
This is an example ruby script from the book content, which seems to
have been brought in with the images long ago in 05cde251 (prepare for
progit v2, 2014-10-23) . But we don't actually read it! We pull the
updated version in as part of a normal book import.

Which is a good thing, because it got corrupted by 454ebb52 (Update code
using rubocop's autocorrect linter, 2018-10-28). Let's delete it, which
avoids confusion, especially when running a more recent rubocop (which
wants to lint it further).