rubocop: disable Style/StringLiterals
This wants to convert all of our double-quotes to single-quotes. Which
isn't wrong or even bad, but AFAIK it's just a style question, and it
doesn't seem worth disrupting the code base.

I think we could also tell it that we prefer double-quotes (which does
seem to be what most of the code base is using), but I'm not sure it's
even worth the effort.