rubocop: ignore style warnings for db schema
This suggests dumb things like treating date numbers like 2012_07_04
to be split at every third character. And it fights with the rails
autogeneration, which puts a blank line at the beginning of the block
(but rubocop really wants to remove it, so you go back and forth as
rails and rubocop are run).