index: make return of new_content clear in expand_content()
The actual return of expand_content() is hidden behind a bunch of
conditionals. Let's make it clear at the end that we are expanding
new_content.

Note that this is technically a little different, because one error code
path would have returned the result of puts. But returning a nil
new_content makes more sense.