rubocop: set Exclude keys to merge
When we added an AllCops.Exclude key in c6736094 (rubocop: ignore
db/migrate/, 2022-10-05), that meant we stopped respecting the built-in
config. And that config contains vendor/**, among other things.

This wasn't a problem for me locally, because my bundled gems are found
elsewhere. But in our CI environment, they are in vendor/, which is what
caused all of the problems with out-dated rubocop config in gems.
Hopefully this fixes it.