index_doc(): loosen match of "git-cmd(1)" on manpages
In v2.38.0, we now have a manpage for scalar, which doesn't start with
"git". As a result, our regex to find the section number fails, and the
whole manpage is aborted.

Let's loosen this regex to match any alphanumeric command name, plus
dashes. That's enough for the current code base. It's conceivable we
might introduce more characters, but you get a nice big exception
if we do.