Convert the Windows downloads page to Hugo
This migrates the page from the Rails app to the static site generated
via Hugo.

We also need to move the server-side logic that is responsible for
adjusting the download link on 32-bit Windows to the client side, as
Javascript.

Note that we need to access the site parameters from that Javascript
code, therefore we need to move the `application.js` file from `static/`
to `assets/`. While at it, also minify it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>