Adjust the "references" page to Hugo
The categories shown on the references page are added to `data/` and
used via the new `category` shortcode.

This new shortcode is actually only a shim that hands off to the new
`category` partial layout; This is done in that manner to allow reusing
the latter in the upcoming commit where the `Topics` drop-down of the
manual pages is populated.

Signed-off-by: Victoria Dye <vdye@github.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>