Turn what used to be the `index.rake` file into a proper Ruby script
In the Hugo world, we won't use Rake.

The thema con variazioni is that the ActiveRecords are replaced by a
"YAML database": instead of `<something>.where(...)` to query for a
given `ActiveRecord`, we're reading the file `data/docs.yml` and look
for a given entry in the parsed YAML object.

Likewise, instead of calling `<something>.save` all the time, we store
the data in that YAML object, which gets written once, at the end, when
all the updates have been done.

This might look like abandoning best practices by replacing a database
with a bunch of Git-tracked, generated static files. The idea, however,
is to turn this entire thing into a static web page, therefore we do not
even _want_ to have a database.

Co-authored-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Victoria Dye <vdye@github.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>