lib/tasks/index.rake: generate scalar documentation
Add 'scalar' to the document-matching regexes in 'index_doc()' and
'index_l10n_doc' so that the 'scalar' command documentation is built and
included on the site.

Note that, because the Scalar technical design document has the same base
name as the command doc (both 'scalar.txt'), only last one processed will be
saved to the DB and rendered on the site. Since we want both documents
available on the website, this should be fixed by renaming the technical
document in Git. Temporarily, this will be handled on the site via a later
commit.

Signed-off-by: Victoria Dye <vdye@github.com>