book2: fix frozen string literal problem
Book generation has been broken since 6a1d3bc8 (freeze string literals
in rake tasks, 2022-10-05). The issue is the use of force_encoding() on
a literal, which doesn't work if it's frozen.

One solution would be to just .dup the literal. However, this
force_encoding is not necessary at all these days. Since ruby 2.0,
strings are UTF8 by default. So we can just get rid of it.