Gemfile: update rails
This clears up some CVEs, some of which are high-severity SQL
issues, including an injection. I don't think we're actually vulnerable,
as the threats are pretty specific and we don't take a lot of user input
(though we do in the form of converting URL paths into SQL for things
like book versions, command names for manpages, etc).

It certainly doesn't hurt to err on the side of being up-to-date on such
things.

Done with `bundle update rails` (the problems are actually in
activerecord, etc, but dependabot is not willing to update
them because it requires doing the whole rails dependency tree).