Implement client-side search using Pagefind
We now use Pagefind (https://pagefind.app/) to implement a client-side,
Javascript-based search.

For now, the "live search" is not implemented, only the full search
results (i.e. what happens after pressing the Enter key in the search
box).

Pagefind is maintained actively, integrates well with Hugo, and is very
fast, in particular because in contrast to most other client-side search
solutions the search index is not one single, gigantic file, but is
downloaded in chunks, as needed.

Alternatives to Pagefind I considered were:

- `lunr.js` (https://lunrjs.com/), one of the oldest client-side search
  solutions. Indeed, the first iteration of client-side search that I
  implemented was using `lunr.js`.

  Sadly, lunr.js seems to be no longer actively maintained. And it comes
  at the price of one big search index file. Even if a pre-processed
  search index is used, even restricting to English and only to the
  newest versions of the manual pages, it weighs in with 3 megabytes,
  which takes a prohibitively long time load while the user is waiting
  for search results.

- https://github.com/jameslittle230/stork

  As of time of writing, this seems to have become unmaintained.

- https://github.com/tinysearch/tinysearch

  Seems to be limited to whole-word search.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>