book: keep colons and other special characters in URLs
In the Korean translation, there is a section whose title contains a
colon. By default, Hugo strips such characters from the URLs it
produces. However, for backward compatibility, the existing URLs in the
current Rails App backing git-scm.com (which do contain colons in these
instances) need to keep working.

Likewise, the Japanese and Azerbaijani translations (and probably
others, too) include a variety of special characters in their page URLs
such as apostrophes and parentheses.

So let's ask Hugo, nicely, to please keep them as part of the URL.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>