book: fix translated labels
When generating the HTML files for the ProGit book's translations, we
want to use localized labels for things like appendices, chapters, etc.

There is already code to retrieve those localized labels. However, it
fails to work, for two reasons:

- AsciiDoctor's default branch changed to `main`, and
- `URI.open` is private by default

Fix both of those issues, and for good measure print a warning if the
localized labels could not be loaded.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>