Add a GitHub workflow to update the translated manual pages regularly
After taking care of the English manual pages, let's also have a
scheduled workflow that takes care of updating the translated manual
pages.

Just like the scheduled workflow that takes care of the original,
English manual pages, here we also need a flag to toggle complete
rebuilds in case the script, the style sheets, or the layouts changed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>