Move the files defining the documentation categories
The files' contents still reflect the Rails App, but they are already
moved to the location where they will become YAML files in the next
commit.

This separation is on purpose: It allowed Git to pick up on upstream
changes while developing the PR that migrates git-scm.com to Hugo.

Note that three file names were adjusted to the name of the
documentation category recorded within the file itself.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>