search: give the manual pages' titles maximal weight
When searching for a Git command, chances are that the user wants the
corresponding manual page. So let's try to present that one first thing
by increasing the `data-pagefind-weight`.

Let's increase the weight to the maximal value (10) for the Git command
that is described in the manual page.

To ensure that the word density does not play a role (so that the
`logos` page does not "win" over the `git log` manual page when
searching for "log"), decrease the default weight to 0.05, as suggested
by Pagefind's author.

For further details, see https://pagefind.app/docs/weighting/.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>