search: reintroduce the categories in the live search
In the existing Rails app, when the user types a search term into the
search box, the results are sorted into the categories "Reference" and
"Book". Let's do the same with the Pagefind-based search.

We do this by specifically adding metadata indicating the category of
manual pages as well as for the sections of the book (for details, see
https://pagefind.app/docs/metadata/). Then, we use that information to
sort the results as they arrive into the correct table row (and
displaying it when necessary, as the empty rows are initially hidden).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>