book(xrefs): maintain question marks that are part of the URL path
We need to be careful to allow for a trailing question mark in the
sections' titles; These need to be (URL-)encoded in the URL and not be
mistaken for the separator between path and GET parameter(s).

While at it, also install redirects for URLs containing a question mark
so that incorrect URLs that contain a literal question mark (which is
interpreted as separator between the URL path and the `GET`
parameter(s)) redirect to the correct ones.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>