book: fix a couple of broken redirects
When falling back to redirecting to to the English ProGit2 book in case
a translation is incomplete and is missing a certain reference, we face
the problem that some references have been renamed in the English
version.

Let's list the known mappings and thereby fix most of the broken links.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>