Add a helper script to run Pagefind
Nominally, `pagefind.js` can only be used on a website.

But with a few hacks, as described in
https://github.com/CloudCannon/pagefind/discussions/175#discussioncomment-9038697,
it can also be run via a node.js script on the command-line.

This script will be invaluable in the CI builds, to guarantee that
certain search results appear on top (such as `docs/git-log` for the
search term "log").

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>