css: disable ligatures in monospace fonts
The choice of the font in app/assets/stylesheets/variables.scss for the
rule:

	$fixed-width-font-family: Courier, monospace;

can make it harder to read some of the text in the documentation, at
least in some browsers in some OSes. This is especially applicable to
the names of the configuration keys and to mentions of the shell
commands, which people would usually edit in a text editor or in a shell
with a monospace font. Fonts used by <code> tags are also affected.

For example, the following ligatures "meld" the letters together:

  - in config key `gc.reflogExpireUnreachable` the ligature is `fl`
  - in config key `color.diff.new` -- `ff`
  - in command name `git config` -- `fi`

Add `font-variant-ligatures: none;` [1] to all usages of variable
`$fixed-width-font-family` to fix this readability issue.

[1] see documentation of this CSS property on MDN:
    https://developer.mozilla.org/en-US/docs/Web/CSS/font-variant-ligatures
Fixes: https://github.com/git/git-scm.com/issues/1817