Fix rendering of clone's/init's synopses
Since git/git@76880f0510c (doc: git-clone: apply new documentation
formatting guidelines, 2024-03-29) and its parent commit, the synopses
of the manual pages of the `git clone` and the `git init` command are
rendered in a different way.

This matters because the changes were clearly only tested with AsciiDoc,
but not with AsciiDoctor. Using the latter, the formatting is all off,
unless the default stylesheet is overridden.

On Git's home page, AsciiDoctor is used to render the manual pages, and
the default stylesheet is overridden.

However, the synopses still look unpleasant to my eye, e.g. because of
the inconsistent background color of the `<code>` blocks.

This commit changes the style of the `<code>` blocks in the synopses so
that they flow better again, visually.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>