search: add shortcuts to focus
This cribs from Racket's documentation site [1], where pressing "s" will
focus the search bar. This is a modified version of [2], which I choose
to receive and use under the MIT license [3], [4], which is compatible
with MIT-LICENSE.txt.

Add a note to the search bar's placeholder: there's a probably a better
way to steer users into this, but "/" is common thanks to Vim and many
websites; "s" is natural (at least to me, having used Racket's docs so
long). While we're here, use a true ellipsis character.

The jQuery features may be using deprecated handlers [5], but the site
appears to be on 13-year old jQuery [6]. Modifications to the jQuery
code from Racket work better on Firefox and allow other document
elements to intercept the event.

[1]: https://docs.racket-lang.org
[2]: https://github.com/racket/scribble/blob/02ecb223c8e0e4dda3c00739338c96dcd1d5bac6/scribble-lib/scribble/scribble-common.js#L145C2-L152C13
[3]: https://github.com/racket/scribble/blob/02ecb223c8e0e4dda3c00739338c96dcd1d5bac6/LICENSE.txt
[4]: https://github.com/racket/racket/blob/240aa086c548440e0b11641c40cf79e393a560a4/racket/src/LICENSE-MIT.txt
[5]: https://api.jquery.com/keyup-shorthand/
[6]: https://blog.jquery.com/2011/11/21/jquery-1-7-1-released/

Co-authored-by: Johannes Schindelin <johannes.schindelin@gmx.de>