search: link to "pretty" URLs
One of GitHub Pages' nice features is that you do not have to append
`.html` to the URLs. For example, the following two URLs refer to the
same page:

- https://git-for-windows.github.io/requirements

- https://git-for-windows.github.io/requirements.html

We use this feature in the migration of Git's home page to a static
website, to maintain backwards-compatibility with the Rails app:
Navigating to /docs/git-config will automatically load
/docs/git-config.html.

Pagefind knows nothing about this, of course, and suggests the full URL
including the `.html` suffix. But that does not look as nice, and makes
it actively harder to copy/paste helpful, "pretty" links.

So let's just go ahead and strip off that unneeded file extension from
the search results' links.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>