playwright: let the UI tests pass with the Rails app
Currently, the Git home page is implemented as a Rails app. For various
reasons, this app behaves a bit different than the upcoming
Hugo/Pagefind-based home page.

Let's document the differences via the UI tests.

Let's also make it easy to test against various sites, via the
environment variable `PLAYWRIGHT_TEST_URL`. When set, it is used instead
of https://git-scm.com/. This enables swift and painless validation of
the Hugo/Pagefind site, as it is now easy to augment the Playwright
tests to test something on the original Rails app and then to verify
without much hassle that the Hugo/Pagefind site behaves the same.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>