playwright: add a regression test for the `git remote renom` issue
A bug in one of the translated manual pages was reported via
https://discord.com/channels/1042895022950994071/1043101606549401661/1248660690676224124
where the command `git remote rename` was translated to `git remote
renom`.

While it _is_ true that the French translation of the noun "rename" is
"renom", in this instance the translation is incorrect because it is a
Git command, and Git will complain about an unknown subcommand when
thusly invoked.

It's a bit tricky to find the source of these contents of git-scm.com:
https://github.com/jnavila/git-html-l10n/blob/master/fr/git-remote.txt,

Or at least it _should_ be, but strangely enough that source does not
have the typo and it has not changed for the last 7 months.

The actual _actual_ source of these git-scm.com contents is
https://github.com/jnavila/git-manpages-l10n/blob/master/po/documentation.fr.po,
from which `git-html-l10n` is auto-generated, but even there, the text
correctly reads `git remote rename`.

Digging a bit deeper, this was fixed in
https://github.com/jnavila/git-manpages-l10n/commit/b435bfd (Translated
using Weblate (French), 2022-04-16) and that fix trickled into
https://github.com/jnavila/git-html-l10n/commit/506285b (update
2022-04-22, 2022-04-22) a week later. But apparently almost 2.5 years
later, the fix did not make it to the Git home page. Most likely the
problem is a glitch somewhere in the cache/database of the Rails app.

Document this particular problem as fixed in the Hugo/Pagefind version
of Git's home page, and as broken on the Rails app version, by
specifically checking for this in the Playwright tests.

Reported-by: emmanuel Hesry <ehesry@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>