playwright: verify that `isRailsApp` is correct
The `isRailsApp` variable tells whether the URL in question refers to
the original Rails App variant of https://git-scm.com/ or to the new
Hugo/Pagefind-based site. The variable is set to true if the URL is
https://git-scm.com/.

However, today we tried for the first time to switch
https://git-scm.com/ to the new site. Therefore, that condition would
have been wrong (even as luck had it, the URL in use was a http one, not
an https one, i.e. the insecure variant, and therefore `isRailsApp` was
correctly set to `false`, but that was just happenstance).

Let's make sure that the value of the `isRailsApp` is correct by looking
for the tell-tale `<meta name="generator" content="Hugo <version>">` tag
that is generated by Hugo but not by the Rails app.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>