ci: skip the scheduled workflows in forked repositories
The idea of the scheduled workflows is to keep the actual
https://git-scm.com/ site up to date. To do that, commits are added to
the branch as needed.

Doing this in fork repositories makes little sense, as these commits
will necessarily diverge from the `gh-pages` branch of
https://github.com/git/git-scm.com/.

So let's not run them in forks.

Note: To allow for improving said workflows, manually triggered runs
will still do the work as intended.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>