ci: upload Playwright tests' results even in PR builds
We do that in all other workflows that run the Playwright tests, and it
is very helpful when diagnosing issues (especially the screenshots that
are stored on failure).

This needs a slight change in the Playwright configuration where we now
want to have the `github` reporter _in addition_ to the `html` reporter
(and not instead of it).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>