book: re-render with the `+` work-around
In https://github.com/git/git-scm.com/pull/1879 it was reported that at
least the English and Korean version of chapter 10 section 5 displayed a
regression where the explanation of the refspec format was formatted
incorrectly.

Turns out that this is a regression in AsciiDoctor v2.0.19. We just
adjusted the `update-book2.rb` script to work around this issue, and now
it is time to commit the result of re-running that script for all
variants of the ProGit book.

This commit is best viewed with `--word-diff`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>