update-book2: work around a regression in AsciiDoctor v2.0.19
In https://github.com/git/git-scm.com/pull/1879 a regression was
reported in the way the explanation of the refspec format is rendered;
It is now broken.

The reason is a regression in AsciiDoctor (reported at
https://github.com/asciidoctor/asciidoctor/issues/4624).

As a work-around, replace the `+` instances (which are mis-interpreted
if more than one exists in the same line) by `{plus}` (which AsciiDoctor
renders exactly like AsciiDoc would).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>