update-book2: do allow the fallback for old image locations to work
Recent revisions of the ProGit book have their images in the `images/`
directory located at the top-level, but older revisions had them next to
the chapters that referenced those images.

The `update-book2.rb` script has a fall-back to allow those images to be
copied from the old-style locations.

This fall-back relies on an exception to be raised when an image does
not show up in the new-style location, but we returned a warning
instead. Let's not. Let's raise that exception again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>