aliases: Keep not only anchor but also query when redirecting
We just fixed a bug where clicking on an operating system button in
https://git-scm.com/downloads/guis replaced the URL by an incorrect URL,
e.g. https://git-scm.com/download/guis?os=linux instead of
https://git-scm.com/downloads/guis?os=linux (note the singular
"download" in the incorrect URL).

We want to handle such URLs gracefully, though, just in case those links
were already shared.

Therefore we are about to add aliases, but the way aliases are currently
handled loses the `?os=linux` part of the URL, which is rather
important.

Let's copy that part (which is called `window.location.search` in
Javascript) into the new URL in the Javascript part of the redirect
page (which is typically taking precedence over the `<meta
http-equiv="refresh" ...>` tag).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>