Fix the sidebar button on iPhones
As per https://webkit.org/b/22261#c68, the Safari browser never gives
buttons the focus.

This breaks the expectation of git-scm.com's sidebar button that is
shown on small screens: clicking on it will never open that sidebar and
navigation is made harder than necessary by that.

The solution, as per the hint at the incredibly helpful post at
https://stackoverflow.com/questions/42758815/safari-focus-event-doesnt-work-on-button-element#53157834
is to turn the <button> into a <div tabindex="1">. That way, it works
even on iPhones. With minor style adjustments, the button even looks the
same as before!

Reported-by: Toon Claes <toon@iotcl.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>