playwright(book): let the tests pass on small screens
When the Playwright tests want to verify that the translated book
versions can be navigated to, they assumed that the links were visible,
but that is only true on "big screens". On phones, for example, the
sidebar button has to be pressed first.

With this change, a manual `npx playwright test` passes again.

Note: This was not caught by the CI runs because they only test with
Chrome (to avoid hefty downloads of the Firefox, Chromium and Webkit
browsers, only Chrome is tested because it comes pre-installed on
GitHub's hosted Actions runners).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>