.github/actions/deploy-to-github-pages: use Cloudflare API directly
As pointed out in [1], using the cloudflare-purge-action incurs a ~31
second penalty at the start of the "deploy" action, where time is spent
building a Docker container to run the action.

This is unnecessary, since Cloudflare has a straightforward REST API
that we can use cURL to communicate with directly, without the extra
start-up cost.

Let's do that instead, and move this to run in the
deploy-to-github-pages action, which is run from multiple entry points,
all of which will want to purge the Cloudflare caches upon deployment.

[1]: https://github.com/git/git-scm.com/pull/1893#issuecomment-2392975994

Signed-off-by: Taylor Blau <me@ttaylorr.com>